requirements.txt
aiofiles==24.1.0
asgiref==3.8.1
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.0
dict-toolbox==5.0.0
distlib==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.16.0
filelock==3.17.0
Jinja2==3.1.6
lazy-object-proxy==1.11.0
MarkupSafe==3.0.2
msgpack==1.1.0
nest-asyncio==1.6.0
paramiko==3.5.0
platformdirs==4.3.6
pop-config==12.0.4
pop-loop==1.1.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
PyYAML==6.0.2
rend==7.0.2
sniffio==1.3.1
sqlparse==0.5.3
toml==0.10.2
virtualenv==20.29.1
